just commit

+++ src/main/java/com/caselab/project/repository/FileRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FileRepo extends JpaRepository<File, Integer> {
+public interface FileRepo extends JpaRepository<File, Long> {
 }
